Convert makeTemporaryDirectories and determineOutputDirectory to regular methods and call them at setup
When resuming, we were losing the output directory info, thus the resulting img was not written in the expected directory.
